Renamed GetVcPlatform back to GetPlatform
